requirements.txt
Quart==0.19.4
aiomysql==0.2.0
ccxt==4.5.21
pandas==2.1.0
numpy==1.26.0
python-dotenv==1.0.0
werkzeug==3.0.0
cryptography
uvicorn